moved chache benchmark tests to its own file and resolved linter warnings
